+++ R/grpmean.R
@@ -20,6 +20,11 @@
 #' @param encoding Character vector, indicating the charset encoding used
 #'   for variable and value labels. Default is \code{"UTF-8"}. Only used
 #'   when \code{out} is not \code{"txt"}.
+#' @param out Character vector, indicating whether the results should be printed
+#'    to console (\code{out = "txt"}) or as HTML-table in the viewer-pane
+#'    (\code{out = "viewer"}) or browser (\code{out = "browser"}), of if the
+#'    results should be plotted (\code{out = "plot"}, only applies to certain
+#'    functions). May be abbreviated.
 #'
 #' @return For non-grouped data frames, \code{grpmean()} returns a data frame with
 #'   following columns: \code{term}, \code{mean}, \code{N}, \code{std.dev},